German Credit Data.

Data from Dr. Hans Hofmann of the University of Hamburg.

data(german)

Format

A data.frame with 1000 rows and 21 variables

Source

UCI Machine Learning Repository https://archive.ics.uci.edu/ml/datasets/Statlog+(German+Credit+Data) http://www1.beuth-hochschule.de/FB_II/reports/Report-2019-004.pdf

Details

These data have two classes for the credit worthiness: Good or Bad. There are predictors related to attributes, such as: checking account status, duration, credit history, purpose of the loan, amount of the loan, savings accounts or bonds, employment duration, Installment rate in percentage of disposable income, personal information, other debtors/guarantors, residence duration, property, age, other installment plans, housing, number of existing credits, job information, Number of people being liable to provide maintenance for, telephone, and foreign worker status.

This is a transformed version of the German Credit data set with factors instead of dummy variables, and corrected as proposed by Grömping, U. (2019).

References

Grömping, U. (2019). South German Credit Data: Correcting a Widely Used Data Set. Report 4/2019, Reports in Mathematics, Physics and Chemistry, Department II, Beuth University of Applied Sciences Berlin.