Get the best submissions per team.

Get the best submissions per team.

get_best(history, metrics = names(metrics), test_name = "quiz",
  decreasing = FALSE)

Arguments

history

list of the submissions history per team as returned by compute_metrics

metrics

character vector. names of the metrics

test_name

string. name of the test set used: "quiz" or "test"

decreasing

logical. Should the sort order be increasing or decreasing? Must be of length 1 or with the same length as metrics.

Value

get_best returns a data.frame where the rows are teams in sorted order of performance. The best submission per team is retained. The sort is based on possibly several metrics in the order given by the metrics argument. In case of ties on the first metric, the second metric is used to break the ties, and so on. Lastly, the date is used in case of ties. The columns are: