Plot the density of submissions over time.

Plot the density of submissions over time.

plot_activity(
  history,
  baseline = "baseline",
  col = 1:length(history),
  alpha.f = 0.7,
  bw = 3600 * 24,
  by = 4,
  xlab = "Date",
  ylab = "Submissions intensity",
  bty = "l",
  fg = "darkslategray",
  col.axis = fg,
  col.lab = fg,
  text.col = fg,
  ...
)

Arguments

history
list of the submissions history per team as returned by compute_metrics

baseline
string. name of the team considered as the baseline that will not be plotted.

col
colors of the teams.

alpha.f
factor modifying the opacity alpha of colors; typically in [0,1].

bw
real. the smoothing bandwidth to be used by density in seconds.

by
real. height of the interval between two teams in number of submissions.

xlab, ylab
axis labels. see title.

bty, fg, col.axis, col.lab
graphical parameters. see par.

text.col
the color used for the legend text. see legend.

...
further parameters passed to plot function.

Value

NULL

See also