Plot the history of the scores of each team over time.

The best score of each team has a bold symbol.

plot_history(
  history,
  metric,
  test_name = "quiz",
  baseline = "baseline",
  col = 1:length(history),
  pch = rep(21:25, 100),
  by = 0.05,
  xlab = "Date",
  ylab = "Score",
  bty = "l",
  fg = "darkslategray",
  col.axis = fg,
  col.lab = fg,
  text.col = fg,
  ...
)

Arguments

history
list of the submissions history per team as returned by compute_metrics

metric
string. name of the metric considered

test_name
string. name of the test set used: "quiz" or "test"

baseline
string. name of the team considered as the baseline. Its best score will be plotted as a constant and will not appear in the legend.

col
colors of the teams

pch
symbols of the teams

by
real. interval width of grid lines

xlab, ylab
axis labels. see title.

bty, fg, col.axis, col.lab
graphical parameters. see par.

text.col
the color used for the legend text. see legend.

...
further parameters passed to plot function.

Value

NULL