rchallenge: A Simple Data Science Challenge System

A simple data science challenge system using R Markdown and Dropbox <https://www.dropbox.com/>. It requires no network configuration, does not depend on external platforms like e.g. Kaggle <https://www.kaggle.com/> and can be easily installed on a personal computer.

Note

The rendering of HTML content provided by Dropbox will be discontinued from the 3rd October 2016 for Basic users and the 1st September 2017 for Pro and Business users. See https://www.dropbox.com/help/16. Alternatively, GitHub Pages provide an easy HTML publishing solution via a simple GitHub repository.

version 1.16 of pandoc fails to fetch font awesome css, see https://github.com/jgm/pandoc/issues/2737.

Installation

Install the R package from CRAN repositories

install.packages("rchallenge")

or install the latest development version from GitHub

# install.packages("devtools")

devtools::install_github("adrtod/rchallenge")

A recent version of pandoc (>= 1.12.3) is also required. See the pandoc installation instructions for details on installing pandoc for your platform.

Getting started

Install a new challenge in Dropbox/mychallenge:

setwd("~/Dropbox/mychallenge")

library(rchallenge)

new_challenge()

or for a french version:

new_challenge(template = "fr")

You will obtain a ready-to-use challenge in the folder Dropbox/mychallenge containing:

challenge.rmd: Template R Markdown script for the webpage.

data: Directory of the data containing data_train and data_test datasets.

submissions: Directory of the submissions. It will contain one subdirectory per team where they can submit their submissions. The subdirectories are shared with Dropbox.

history: Directory where the submissions history is stored.

The default challenge provided is a binary classification problem on the German Credit data set.

You can easily customize the challenge in two ways:

During the creation of the challenge: by using the options of the new_challenge function.

After the creation of the challenge: by manually replacing the data files in the data subdirectory and the baseline predictions in submissions/baseline and by customizing the template challenge.rmd as needed.

Next steps

To complete the installation:

Create and share subdirectories in submissions for each team:

new_team("team_foo", "team_bar")

Render the HTML page: publish() Use the output_dir argument to change the output directory. Make sure the output HTML file is rendered, e.g. using GitHub Pages.

Give the URL to your HTML file to the participants.

Refresh the webpage by repeating step 2 on a regular basis. See below for automating this step.

From now on, a fully autonomous challenge system is set up requiring no further administration. With each update, the program automatically performs the following tasks using the functions available in our package:

store_new_submissions: Reads submitted files and save new files in the history.

print_readerr: Displays any read errors.

compute_metrics: Calculates the scores for each submission in the history.

get_best: Gets the highest score per team.

print_leaderboard: Displays the leaderboard.

plot_history: Plots a chart of score evolution per team.

plot_activity: Plots a chart of activity per team.

Automating the updates on Unix/OSX

For the step 4, you can setup the following line to your crontab using crontab -e (mind the quotes):

0 * * * * Rscript -e 'rchallenge::publish("~/Dropbox/mychallenge/challenge.rmd")'

This will render a HTML webpage every hour. Use the output_dir argument to change the output directory.

If your challenge is hosted on a Github repository you can automate the push:

0 * * * * cd ~/Dropbox/mychallenge && Rscript -e 'rchallenge::publish()' && git commit -m "update html" index.html && git push

You might have to add the path to Rscript and pandoc at the beginning of your crontab:

PATH=/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin

Depending on your system or pandoc version you might also have to explicitly add the encoding option to the command:

0 * * * * Rscript -e 'rchallenge::publish("~/Dropbox/mychallenge/challenge.rmd", encoding = "utf8")'

Automating the updates on Windows

You can use the Task Scheduler to create a new task with a Start a program action with the settings (mind the quotes):

Program/script: Rscript.exe

options: -e rchallenge::publish('~/Dropbox/mychallenge/challenge.rmd')

Examples

Credit approval (in french) by Adrien Todeschini (Bordeaux).

Spam filter (in french) by Marie Chavent (Bordeaux).

Please contact me to add yours.

See also